Split database methods into user and book files
